% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/divclust.R
\name{dogs}
\alias{dogs}
\title{Breeds of Dogs data}
\format{A data frame with 27 rows (the breeds of dogs) and  7 columns: their size, weight and speed with 3 categories
(small, medium, large), their intelligence (low, medium, high), their affectivity and aggressiveness
with 3 categories (low, high), their function (utility, compagny, hunting).}
\source{
Originated by A. Brefort (1982) and cited in Saporta G. (2011).
}
\description{
Data refering to 27 breeds of dogs.
}

