\name{INSEE}

\alias{INSEE}

\docType{data}

\title{INSEE}

\description{Description of 302 cities by 56 variables structured in 8 groups}

\format{
A data frame with 302 observations on 56 variables
}
  
  \usage{data(INSEE)}

  \source{
  www.insee.fr
  }
  
  
  \keyword{datasets}
  