\name{print.PCAmix}
\alias{print.PCAmix}
\title{
Print a 'PCAmix' object
}
\description{
This is a method for the function print for objects of the class \code{PCAmix}.
}
\usage{
\method{print}{PCAmix}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{x}{an object of class  \code{PCAmix} generated by the functions \code{\link{PCAmix}} and \code{\link{PCArot}}.}
  \item{\dots}{further arguments to be passed to or from other methods. They are ignored in this function.
}
}


\seealso{
\code{\link{PCAmix}} , \code{\link{PCArot}}
}

\keyword{ print }