\name{sol.2dim}
\alias{sol.2dim}
\title{Varimax rotation for PCAmix in two dimensions}
\usage{
  sol.2dim(A, indexj, p, p1)
}
\arguments{
  \item{A}{matrix of loadings}

  \item{indexj}{a vector with the number of the variable
  associated with each row of A}

  \item{p}{the total number of variables}

  \item{p1}{the number of quantitative variables}
}
\value{
  \item{theta}{the angle of rotation} \item{T}{the matrix
  of rotation}
}
\description{
  Varimax rotation for PCAmix in two dimensions
}

