% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supvar.PCAmix.R
\name{supvar}
\alias{supvar}
\title{Supplementary variables projection}
\usage{
supvar(obj, ...)
}
\arguments{
\item{obj}{an object of class \code{PCAmix} or \code{MFAmix}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
\code{supvar} is a generic function for adding supplementary variables
in \code{PCAmix} or \code{MFAmix}. The function invokes invokes two methods which depend on
the class of the first argument.
}
\details{
This generic function has two methods \code{\link{supvar.PCAmix}} and
\code{\link{supvar.MFAmix}}
}
