\name{wine}

\alias{wine}

\docType{data}

\title{Wine}

\description{
The data used here refer to 21 wines of Val de Loire.

}

\usage{data(wine)}

\format{
A data frame with 21 rows (the number of wines) and 31 columns: the first column corresponds to the label of origin,
the second column corresponds to the soil, and the others correspond to sensory
descriptors.
}

\source{
Centre de recherche INRA d'Angers
}


\keyword{datasets}
