\name{print.MFAmix}
\alias{print.MFAmix}
\title{
  Print a 'MFAmix' object
}
\description{
  This is a method for the function print for objects of the class \code{MFAmix}.
}
\usage{
  \method{print}{MFAmix}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object of class  \code{MFAmix} generated by the function \code{\link{PCAmix}}. 
  }
  \item{\dots}{
    Further arguments to be passed to or from other methods. They are ignored in this function.
  }
}
\author{Marie Chavent <marie.chavent@u-bordeaux2.fr>, Vanessa Kuentz, Benoit Liquet, Jerome Saracco}

\seealso{
  \code{\link{MFAmix}}
}

\keyword{ print }