\name{wine}

\alias{wine}

\docType{data}

\title{Wine}

\description{
The data used here refer to 21 wines of Val de Loire.

}

\usage{data(wine)}

\format{
A data frame with 21 rows (the number of wines) and 31 columns: the first column corresponds to the label of origin,
the second column corresponds to the soil, and the others correspond to sensory
descriptors.
}

\source{
Centre de recherche INRA d'Angers

Le, S., Josse, J. & Husson, F. (2008). FactoMineR: An R Package for Multivariate Analysis. Journal of Statistical Software. 25(1). pp. 1-18.
}


\keyword{datasets}
